Major bugfix release
Lots of bugs/issues fixed in this one. All relating to the rat-nest of
code called check_song_loop. You can now set/search for songs before
connecting without it resetting to blank every 5 seconds. Errors that
happened when using foobar in non-default track progression mode (e.g.
shuffle, random, etc) are now gone. Added code to be sure no
funny-business happens with respect to the current song, manually set
song, and whether or not we are connected/disconnected.

My next goal is to multithread the lyrics getter functions. This has to
be done with QThread I think though so its not as easy as you might
think.