bLyrics 2.0 Beta 1
As with all things in my life, I got bored Tuesday and decided to rewrite my old bLyrics application so that it wasn't terrible. Unfortunately I failed in that respect because at best I detangled maybe half the spaghetti code.

Most significantly is that I removed all parts that dealt with the COM interface and instead im pulling all data from the ajquery web interface. This is so much faster, is more reliable, and is far less code.

This program now only requires the foobar2000 web interface plugin with the ajquery template to function.

Manual song selection/song searching is still not working correctly and there may potentially be some errors with unicode stuff.

Some ideas I may implement would be caching of lyrics and the ability to manually set a song's lyrics for songs with incorrect or missing lyrics.