Unicode fix
Ran into the potential unicode issues almost immediately, this update fixes it for the most part. Also made sure that the data grabbing didn't change the foobar2000 interface's current playlist/page/selection. Was really annoying not being able to browse my foobar library with the app open as it always switched back to the current playlist. Fixed now although I think I can do better if I figure out the jquery stuff for the library. Then I wont need to mess with the interface at all. The documentation sucks though so I'm not having a great time of it.

Finally I removed the hardcoded cookie since it appears not to be necessary anymore.