Cleaned up json
When foobar2000 wasn't focused on the currently playing playlist the json queries were causing disruptions to the UI that was extremely annoying. I tried to minimize disruptions to only when necessary but it still makes browsing the library nearly impossible. I think the only solution is just to cut this all out and accept that some times we won't get data on the next song.