Increased reliability
Now the app can retrieve the currently playing song 99% of the time without issue and can get the next song as long as the current playlist is focused and the current song isnt on the last page of the web interface.

Just found some actual docs on the http component of foobar2000 but it doesn't appear to have any pure library querying functions that don't also affect the actual fb2k UI. I did find some interesting functions for getting the current queue which could be useful. If the user has a queue set then we can get the next song regardless of the currently focused playlist/page.