Removed threading
Was a nice idea but combining vanilla threading with Qt is fraught with issues I don't want to deal with right now. It was really unnecessary anyway since the web interface should be local most of the time. The part that needs the threading is the lyrics snatching but that will definitely have to use proper QThreads.

In the end this version is still so much faster than the old COM version.