Fix stopped song status
Previously when stopped it would use whatever song was currently highlighted as the current song. This is only true when foobar2k first starts (i.e. prevplayedItem == ""). When prevplayedItem data is available we now use this instead which accurately reflects which song will actually be played when the player is started.