Lyrics caching and more UTF fixes
Implemented basic lyrics caching. It saves the lyrics in a file named using a base64 encoded version of the song and artist names. Its not easy to search through as a human but is far easier than dealing with special chars in filenames. Next feature will be an easy way to modify the lyrics through the GUI.

I also think I implemented a far better fix for any utf issues. Really I should have done this first off but I didn't. I just moved the encode() calls to the places where we get the data for the very first time (inside foobarStatusDownloader).

There is now an option in the File menu to clear the cached lyrics files. You can also delete the .lyricscache folder in the same directory as this script.