Many fixes
First and foremost I fixed the lyrics cacher which completely broke after the modularization update. Fixed a bug which prevented lyrics from being cached if the cache directory was deleted during runtime. Fixed the songlyrics.com source to properly return None when the lyrics aren't on the site. Also removed a bunch of old code I accidentally left in lyricswiki.py. Finally fixed a bug which prevented edited lyrics from being properly saved. It was saving them, but it was using the wrong filename (song name was passed as the artist name on accident).