Basic QThread implementation
I got a very basic threaded implementation for the lyrics downloading functions using QThreads. It wasn't as difficult as I thought it would be but there's a lot of things to be aware of. We basically can't interact with Qt from inside the thread without crashing all over the place. Have to use signals/slots to get anything done. Even something as simple as printing doesn't work because my print function update a QTextDocument, which you cant do from a different thread. Have to rework the app around signals/slots to fully impement things.

The cache builder is completely broken by this update and has to be reworked a bit to work with threads properly.