Multithreading Finished
Got the multithreading all working now without any stack limit issues. I realized in the middle of rewriting the cachebuilder (for the second time) that I had originally intended to do it this way in the first place. Dunno where I got off track. I've used it to pregenerate my own playlist of ~1200 songs without issue.

Next up is to add modules for both Genius and AZlyrics as they seem to be quite up to date compared to songlyrics and lyricswiki.