Bug fixes
Removed all of the manual query/search dialog stuff since it wasn't functional anyway. I do plan on reimplementing those features but in a completely different way.

Fixed a bug which caused the lyrics cache folder to be saved in a temporary folder when running from an executable. Fixed for pyinstaller-based executables but not for py2exe (which refuses to run on my machine).