Genius lyrics provider
Whipped up a quick and dirty Genius lyrics grabber. You are supposed to use their API with a key but I cba to set a key for everyone who uses this app so this is fine for now. If it ever breaks it wont break the entire app, we just fallback to other sources. Keeping up with the changes is going to be very difficult though. I have no way of knowing if things break other than just not getting lyrics from that provider anymore. Might have to reconsider ways to exfiltrate error information from the lyrics providers. At the moment they just return None on erroring.

Also fixed a bug in the newly-minted azlyrics provider. Never a dull moment when I'm programmin...