Console output bug
Last version would slowly degrade in performance and eventually crash the longer the console output got to be. I had forgotten when I first made this program I designed the console output not in the smartest way. I was tracking the output in a list and clearing/resetting the text of the console every update. This is fine when the console isnt thousands of lines long. When the cursor/scrollbar manipulation was added it exacerbated the problem to where it would crash after less than a hundred or so lines of console output.

The solution was to just insert the new text at the bottom without updating anything else. Would have done that in the first place but I hadn't learned that lesson until now.