Search finished!
Wow, I honestly thought this would be harder than it was. Every step of the way was so much easier than anticipated, no issues whatsoever.

Can now search your cached lyrics for song names, artist names, and lyrics strings. If you know a part of a song but dont know the artist or song name you can just type that bit in and find all songs with that specific string in them.

Now to think of the next feature...