added search progress indicator
Added a progress indicator to the search button so that the user can see the current progress of the search. The search button now displays the total number of cached lyrics and the current file number it is searching (i.e. 123/456). Previously there was no indication of progress.

Also moved the call to processEvents to the beginning of each search loop iteration. Previously the call would be skipped on files that didn't match the search parameters.