code separation
In preperation for adding new source for foobar2k's status I decided to reorganize the code to make that task easier. Unfortunately I discovered the new information source was not as verbose as necessary all functions to work properly. The separation is still a good idea however and maybe another source will come along.