Lyrics sources in options
Im adding the ability to manage the lyrics sources from within the application. For the first phase Ive added a new tab in the options dialog with a list widget that shows the current sources along with their version and file name (with path if you mouseover). Dragging will eventually change the source priority. The new priorities will then be saved in the apps preferences and loaded next time to preserve source order across application restarts.

The next goal is to allow a user to refetch lyrics using a specific selection/ordering of sources they select. Some songs just arent correct on even the most reliable sources so you need to be able to try different ones. Currently, the first source to return lyrics is always used, even if those lyrics are wrong.