Reworked data structures
The data structures I was using for the lyrics provider list in the options dialog were abysmal. I've reworked those structures to be nice and consistent all around using nicely formatted dictionaries instead of lists and dicts with weird concatenated key names.

Now to get blyrics to honor the custom priorities is next on the task list.