Bug fix + usability improvement
Fixed a bug that prevented launching the application when there weren't already settings for lyrics source priorities saved. The fallback default were horribly broken.

Added a highlight to the currently selected row in the lyrics source list in the options dialog.

Also a bunch of minor cleanups here and there.