Removed case sensitivity from song search
In the search pane, when the exact checkbox is checked for the song title, it now ignores case differences to allow greater matching. Before if the entered song title was off even by just a capitalization difference, it wouldn't match. This change is in-line with how the lyrics and artist entries work.