Obscure bugfix
And the award for obscure-bug-of-the-month goes to...... This damn thing!

So to get to this bug you have to have the cache loaded into memory, play a song bLyrics can't find lyrics for, and that song must be from an artist that doesn't have any other saved lyrics (not in our cache). Then you have to try and edit that song's lyrics and save it. Then, and only then, do you run into this bug. The fact that I ran into this bug at all is pretty amazing. I think I've seen it once or twice before but never investigated further when I failed to replicate it.

Maybe I should try and clean up some of this nearly-dead code that gets called once in a blue moon. The saveLyrics() method could easily be replaced with whatever routine is saving lyrics after they are downloaded, or vice-cersa. Either way we are saving lyrics using two different systems of code, and this one is pretty much disused.