Basic SQLite support
Implemented the basics for sqlite support. It can create a database and populate it with lyrics. It also currently still works with manual saving and editing. The search dialog is broken still and occassionally it fails to pull lyrics from any provider unless you manully tell it to which is weird (returns the status 'lyrics not found on providers: x,y,z,etc' but manually running on x,y, or z returns lyrics).

The convert option in the Help menu will have proper dialog with progress bar when done.