Cache conversion dialog
Added a dialog to show the progress of the cache conversion. The numbers reported in parenthesis are the number of new songs added to the database out of the total number of cache files. Switched the primary keys back to song+artist in the database since the id idea was bad.