Added library tab
Wasn't going to add anything new before I made a new release but I really wanted to get the library tab into this one. In addition, now when searching for lyrics the active search tab will update with the number of results. Before it would only update on completion. Finally, I update the spec file and the readme for building the single exe file to reflect the removal of suds code.