provider update
Just realized that AZlyrics and Genius were broken again, this time AZlyrics was hiding their content with CSS that I couldn't figure out how to bypass with normal python Requests so I went overkill with Selenium.

I HATE Selenium. Its painfully slow and resource intensive but it does get the job done. I'm using the Firefox webdriver just because that's the browser I have installed but I am very curious if the Chrome driver is as slow. I made the same changes to the Genius provider since it seemed like the same issue. I honestly don't know, I just changed the requests parts to the new selenium code and it started working again so it wasn't my regular expressions (they didnt change the html, the usual reason for a provider breaking).

The other option I am considering now is just pulling the lyrics data from Google search. That wasn't possible until recently, when they decided to add a special header for lyrics search results with the actual lyrics. It seems pretty comprehensive but it pulls from a lot of different sources than we use. Worth looking into because this current "solution" is abhorrently slow, resource intensive, and I hate it.