Updated fail mode
Instead of just reporting the failure to get lyrics, it will not also save that failure as the lyrics for that song. That way the next time the song is played it won't try and download lyrics for the song, it will just display the error message with the date and time of the last download attempt.

Probably should have done this a long time ago.