Genius update
Trying to get Genius to work with PhantomJS but somehow they are detecting it and returning different results (seems to be a page that contains editing options but no lyrics). I might have to wait longer for the page to load using WebDriverWait() and selenium.webdriver.support.expected_conditions. Using the geckodriver does work now though, although its slow and pops a window up (which only seems to happen in the bundled pyinstaller-based exe).